move all logic regarding prices to utils, only export getDisplayPrices and use that on pages
